requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==37.0.4
Django==3.2.7
django-appconf==1.0.5
django-cors-headers==3.8.0
django-crontab==0.7.1
django-cryptography==1.1
django-environ==0.7.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
ez-setup==0.9
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
phonenumbers==8.12.52
Pillow==9.1.1
pycparser==2.21
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.2,
django-cors-headers
boto3>=1.24.82,<1.25
pandas~=1.4.4